backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
statsmodels==0.14.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.5.0
firebase-admin==6.2.0
plotly==5.15.0
joblib==1.3.1
python-dateutil==2.8.2 